fotocalorica_repo_full/backend/requirements.txt
fastapi
uvicorn[standard]
pillow
transformers>=4.30.0
torch
numpy
